Move ExactSizeIterator to own module
